+++ src/background/commands.js
@@ -1,5 +1,20 @@
 "use strict";
-import { BROWSER } from "../core/constants.js";
+import {
+	ALL_CMD_KEYS,
+	BROWSER,
+	CMD_AND_CXM_MAP_TO_WHAT,
+	CMD_EXPORT_ALL,
+	CMD_OPEN_SETTINGS,
+	SETUP_LIGHTNING_PATTERN,
+	TOAST_WARNING,
+} from "../core/constants.js";
+import { openSettingsPage } from "../core/functions.js";
+import Tab from "../core/tab.js";
+import {
+	bg_getCurrentBrowserTab,
+	bg_notify,
+	checkLaunchExport,
+} from "./utils.js";
 
 /**
  * Retrieves all or specified command shortcuts available in the browser extension.
@@ -32,4 +47,40 @@ export async function bg_getCommandLinks(
 	);
 	callback?.(requestedCommands);
 	return requestedCommands;
+}
+
+/**
+ * Listens for extension command events and executes appropriate actions
+ * based on the current Salesforce Setup page context and command received.
+ */
+export function listenToExtensionCommands() {
+	BROWSER.commands.onCommand.addListener(async (command) => {
+		// check the current page is Salesforce Setup
+		const browserTabUrl = (await bg_getCurrentBrowserTab())?.url;
+		if (!browserTabUrl?.match(SETUP_LIGHTNING_PATTERN)) { // we're not in Salesforce Setup
+			return;
+		}
+		const message = {
+			what: CMD_AND_CXM_MAP_TO_WHAT[command] ?? command,
+			url: Tab.minifyURL(browserTabUrl),
+			org: Tab.extractOrgName(browserTabUrl),
+		};
+		switch (command) {
+			case CMD_OPEN_SETTINGS:
+				openSettingsPage();
+				return;
+			case CMD_EXPORT_ALL:
+				if (!checkLaunchExport(undefined, true)) {
+					return;
+				}
+				break;
+			default:
+				if (!ALL_CMD_KEYS.has(command)) {
+					message.what = TOAST_WARNING;
+					message.message = `Received unknown command: ${command}`;
+				}
+				break;
+		}
+		bg_notify(message);
+	});
 }
